Moved test files under __tests__ folder for state folder
