+++ app/state/reducers/__tests__/dataReducer.spec.js
@@ -1,11 +1,11 @@
 import { createAction } from 'redux-actions';
 import Immutable from 'seamless-immutable';
 
-import types from '../../constants/ActionTypes';
-import Reservation from '../../utils/fixtures/Reservation';
-import Resource from '../../utils/fixtures/Resource';
-import User from '../../utils/fixtures/User';
-import dataReducer, { handleData } from './dataReducer';
+import types from '../../../constants/ActionTypes';
+import Reservation from '../../../utils/fixtures/Reservation';
+import Resource from '../../../utils/fixtures/Resource';
+import User from '../../../utils/fixtures/User';
+import dataReducer, { handleData } from '../dataReducer';
 
 describe('state/reducers/dataReducer', () => {
   describe('initial state', () => {
