+++ app/state/reducers/__tests__/notificationsReducer.spec.js
@@ -1,9 +1,9 @@
 import Immutable from 'seamless-immutable';
 import { createAction } from 'redux-actions';
 
-import types from '../../constants/ActionTypes';
-import constants from '../../constants/AppConstants';
-import notificationsReducer from './notificationsReducer';
+import types from '../../../constants/ActionTypes';
+import constants from '../../../constants/AppConstants';
+import notificationsReducer from '../notificationsReducer';
 
 describe('state/reducers/notificationReducer', () => {
   describe('initial state', () => {
