+++ app/state/reducers/api/__tests__/fetchCountsReducer.spec.js
@@ -1,8 +1,8 @@
 import { createAction } from 'redux-actions';
 import Immutable from 'seamless-immutable';
 
-import types from '../../../constants/ActionTypes';
-import fetchCountsReducer from './fetchCountsReducer';
+import types from '../../../../constants/ActionTypes';
+import fetchCountsReducer from '../fetchCountsReducer';
 
 describe('state/reducers/api/fetchCountsReducer', () => {
   describe('initial state', () => {
