+++ app/state/reducers/ui/__tests__/modalsReducer.spec.js
@@ -1,9 +1,9 @@
 import { createAction } from 'redux-actions';
 import Immutable from 'seamless-immutable';
 
-import types from '../../../constants/ActionTypes';
-import ModalTypes from '../../../constants/ModalTypes';
-import modalsReducer from './modalsReducer';
+import types from '../../../../constants/ActionTypes';
+import ModalTypes from '../../../../constants/ModalTypes';
+import modalsReducer from '../modalsReducer';
 
 describe('state/reducers/ui/modalsReducer', () => {
   describe('initial state', () => {
