+++ app/state/reducers/ui/__tests__/reservationsReducer.spec.js
@@ -3,8 +3,8 @@ import Immutable from 'seamless-immutable';
 import first from 'lodash/first';
 import last from 'lodash/last';
 
-import types from '../../../constants/ActionTypes';
-import { DEFAULT_SLOT_SIZE } from '../../../constants/SlotConstants';
+import types from '../../../../constants/ActionTypes';
+import { DEFAULT_SLOT_SIZE } from '../../../../constants/SlotConstants';
 import {
   cancelReservationEdit,
   changeAdminReservationFilters,
@@ -18,10 +18,10 @@ import {
   selectReservationToEdit,
   selectReservationToShow,
   toggleTimeSlot,
-} from '../../../actions/uiActions';
-import Reservation from '../../../utils/fixtures/Reservation';
-import { getTimeSlots } from '../../../utils/timeUtils';
-import reservationsReducer from './reservationsReducer';
+} from '../../../../actions/uiActions';
+import Reservation from '../../../../utils/fixtures/Reservation';
+import { getTimeSlots } from '../../../../utils/timeUtils';
+import reservationsReducer from '../reservationsReducer';
 
 describe('state/reducers/ui/reservationsReducer', () => {
   describe('initial state', () => {
