+++ app/state/reducers/ui/__tests__/resourceMapReducer.spec.js
@@ -2,11 +2,11 @@ import keyBy from 'lodash/keyBy';
 import { createAction } from 'redux-actions';
 import Immutable from 'seamless-immutable';
 
-import types from '../../../constants/ActionTypes';
-import { toggleResourceMap } from '../../../actions/uiActions';
-import Resource from '../../../utils/fixtures/Resource';
-import Unit from '../../../utils/fixtures/Unit';
-import resourceMapReducer from './resourceMapReducer';
+import types from '../../../../constants/ActionTypes';
+import { toggleResourceMap } from '../../../../actions/uiActions';
+import Resource from '../../../../utils/fixtures/Resource';
+import Unit from '../../../../utils/fixtures/Unit';
+import resourceMapReducer from '../resourceMapReducer';
 
 describe('state/reducers/ui/resourceMapReducer', () => {
   describe('initial state', () => {
