+++ app/state/reducers/ui/__tests__/searchReducer.spec.js
@@ -2,10 +2,10 @@ import keyBy from 'lodash/keyBy';
 import { createAction } from 'redux-actions';
 import Immutable from 'seamless-immutable';
 
-import types from '../../../constants/ActionTypes';
-import { clearSearchResults, selectUnit, toggleMap } from '../../../actions/searchActions';
-import Resource from '../../../utils/fixtures/Resource';
-import searchReducer from './searchReducer';
+import types from '../../../../constants/ActionTypes';
+import { clearSearchResults, selectUnit, toggleMap } from '../../../../actions/searchActions';
+import Resource from '../../../../utils/fixtures/Resource';
+import searchReducer from '../searchReducer';
 
 describe('state/reducers/ui/searchReducer', () => {
   describe('initial state', () => {
